pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx-sse>=0.4.3",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.2",
]
